refactor(local-community): extract comment-updates helpers to local-community/comment-updates.ts
Moves _calculateLocalMfsPathForCommentUpdate, _calculateNewCommentUpdate,
_validateCommentUpdateSignature, _updateCommentsThatNeedToBeUpdated,
_syncPostUpdatesWithIpfs, _adjustPostUpdatesBucketsIfNeeded, and
pubsubTopicWithfallback into a focused module. Drops 'private' from
_pageGenerator, _postUpdatesBuckets, and _communityUpdateTrigger so the new
free functions can access them via the community instance. Updates pubsub.ts
and cleanup.ts to call the free helpers directly instead of bouncing through
the class.
